requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==4.4.2
distlib==0.3.8
easygui==0.98.3
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
future==0.18.3
h11==0.14.0
huepy==1.2.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instabot==0.117.0
itsdangerous==2.1.2
Jinja2==3.1.3
ldap3==2.9.1
MarkupSafe==2.1.5
mock==5.1.0
moviepy==1.0.3
numpy==1.26.3
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pdfkit==1.0.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
proglog==0.1.10
psycopg2-binary==2.9.9
pyad==0.6.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pylnk3==0.4.2
PyNaCl==1.5.0
PySocks==1.7.1
pyspnego==0.10.2
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinrm==0.4.3
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
requests-ntlm==1.2.0
requests-toolbelt==1.0.0
responses==0.25.0
schedule==1.2.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sspilib==0.1.0
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.25.1
Werkzeug==3.0.1
winshell==0.6
xmltodict==0.13.0
